requirements.txt
py-cord==1.7.3
python-dotenv==0.19.2
blagues-api==1.0.1
psycopg2==2.9.3